# openapi.yml
openapi: 3.0.0
info:
  title: Jobs API
  version: 1.0.0
  description: API for managing job listings
paths:
  /jobs:
    get:
      summary: Get all job listings
      responses:
        '200':
          description: A list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
    post:
      summary: Create a new job listing
      requestBody:
        description: Job object that needs to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
      responses:
        '201':
          description: Job created
  /jobs/{id}:
    get:
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
    put:
      summary: Update a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Job object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
      responses:
        '200':
          description: Job updated
    delete:
      summary: Delete a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job deleted
components:
  schemas:
    Job:
      type: object
      required:
        - title
        - description
        - company
        - location
        - salary
      properties:
        title:
          type: string
        description:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: number
        datePosted:
          type: string
          format: date-time